+++ guillotina_elasticsearch/events.py
@@ -25,7 +25,6 @@ class SearchDoneEvent(object):
 
 
 class IIndexProgress(Interface):
-    request = Attribute('Request responsible')
     context = Attribute("Context where the Index process has started")
     processed = Attribute('Docs finished indexing')
     total = Attribute('Amount of docs to index')
@@ -35,8 +34,7 @@ class IIndexProgress(Interface):
 @implementer(IIndexProgress)
 class IndexProgress(object):
 
-    def __init__(self, request, context, processed, total, completed=None):
-        self.request = request
+    def __init__(self, context, processed, total, completed=None):
         self.context = context
         self.processed = processed
         self.total = total